JSON Parser
Fairly simple code, I built a dashboard for our CEO and needed to convert some SQL to JSON. A few of the graphs were intensive so to better utilize CPU cycles I implemented the c# part to only look at the last line so know if it needs to jump in or out of the current child. Originally it had more levels so the current one is a simpler form.