Image Screen Scraping
Here was an example of logging into a site and spoofing a client login to get to a protected page systematically, and the save a image every 10 minutes to our server. A quick project as I have been building an entire eCommerce site for my company in the last few months.