OCR Fax to Order
My Company wanted a fast way to convert faxes to orders in our Navision system. I wrote this for them to take the images emailed from our fax (.tif files) and run them with an opensource OCR, then using our public item info I created rules and ranks to find what item a given customer is trying to order. There are also two test images and output here. Once it is done a sales member receives the .csv which they compare to the fax and then uploads it. I have had ~90% correct guesses of cataloged items with this current version.